Fix database to work with new interface
