+++ src/main/kotlin/io/github/kr8gz/egghunt/world/EggFindDetector.kt
@@ -1,7 +1,7 @@
 package io.github.kr8gz.egghunt.world
 
-import io.github.kr8gz.egghunt.Database.foundEgg
-import io.github.kr8gz.egghunt.Egg
+import io.github.kr8gz.egghunt.Database
+import io.github.kr8gz.egghunt.Database.checkAndFindEgg
 import io.github.kr8gz.egghunt.config.config
 import io.github.kr8gz.egghunt.eggHuntMessage
 import net.fabricmc.fabric.api.event.player.AttackBlockCallback
@@ -35,22 +35,19 @@ object EggFindDetector {
     }
 
     private fun eggFindListener(player: PlayerEntity, world: World, pos: BlockPos) {
-        val egg = Egg.findAtLocation(pos) ?: return
-
-        if (!player.foundEgg(egg)) {
-            player.eggHuntMessage("You already found this egg!", Formatting.RED)
-            return
-        }
-
-        player.eggHuntMessage("You found an egg!", Formatting.GREEN)
-
-        with(config.onEggFound) {
-            if (spawnFireworks) spawnFirework(world, pos)
-            world.server?.run {
-                commands.forEach {
-                    commandManager.executeWithPrefix(commandSource, "execute as ${player.uuid} run $it")
+        if (player.checkAndFindEgg(pos)) {
+            player.eggHuntMessage("You found an egg!", Formatting.GREEN)
+            with(config.onEggFound) {
+                if (spawnFireworks) spawnFirework(world, pos)
+                world.server?.run {
+                    commands.forEach {
+                        commandManager.executeWithPrefix(commandSource, "execute as ${player.uuid} run $it")
+                    }
                 }
             }
+        } else {
+            player.eggHuntMessage("You already found this egg!", Formatting.RED)
+            return
         }
     }
